pyproject.toml
# payra-sdk-python/pyproject.toml

[project]
name = "payra_sdk"
version = "1.2.6"
description = "Python SDK for Payra payment signature generation (backend)"
readme = "README.md"
authors = [{ name = "Wraith", email = "support@payra.cash" }]
license = { file = "LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "web3>=6.0.0",
    "eth-abi>=4.0.0",
    "eth-account>=0.10.0", # eth-account depends on eth-utils, so eth-utils is implicitly installed
    "python-dotenv>=1.0.0", # For example usage
]

[project.urls]
"Homepage" = "https://github.com/payracash"
"Source" = "https://github.com/payracash/payra-sdk-python"
"Issues" = "https://github.com/payracash/payra-sdk-python/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["payra_sdk"]
